requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.31.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==22.0.0
python-dotenv==1.0.1
lxml==5.2.2
groq==0.9.0
